+++ subprojects/library/src/commonSample/kotlin/org/kotools/types/ZeroCompanionCommonSample.kt
@@ -2,6 +2,7 @@ package org.kotools.types
 
 import kotools.types.experimental.ExperimentalKotoolsTypesApi
 import kotlin.test.Test
+import kotlin.test.assertNotNull
 import kotlin.test.assertTrue
 
 @OptIn(ExperimentalKotoolsTypesApi::class)
@@ -16,5 +17,12 @@ class ZeroCompanionCommonSample {
             false
         }
         assertTrue(isSuccess)
+    }
+
+    @Test
+    fun fromByteOrNull() {
+        val number: Byte = 0
+        val zero: Zero? = Zero.fromByteOrNull(number)
+        assertNotNull(zero)
     }
 }