+++ subprojects/library/src/commonTest/kotlin/org/kotools/types/ZeroCompanionTest.kt
@@ -5,6 +5,8 @@ import org.kotools.types.internal.InvalidZero
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
+import kotlin.test.assertNotNull
+import kotlin.test.assertNull
 
 @OptIn(ExperimentalKotoolsTypesApi::class)
 class ZeroCompanionTest {
@@ -27,5 +29,22 @@ class ZeroCompanionTest {
         val expected: String = InvalidZero(number)
             .toString()
         assertEquals(expected, actual)
+    }
+
+    @Test
+    fun fromByteOrNullShouldPassWithByteThatEqualsZero() {
+        val number: Byte = 0
+        val actual: Zero? = Zero.fromByteOrNull(number)
+        assertNotNull(actual)
+    }
+
+    @Test
+    fun fromByteOrNullShouldFailWithByteOtherThanZero() {
+        val number: Byte = setOf(Byte.MIN_VALUE..-1, 1..Byte.MAX_VALUE)
+            .random()
+            .random()
+            .toByte()
+        val actual: Zero? = Zero.fromByteOrNull(number)
+        assertNull(actual)
     }
 }